Updated some missed types that should've been update to long with previous check in

+++ src/main/java/com/napier/sem/queries/WorldQueries.java
@@ -217,10 +217,10 @@ public class WorldQueries {
 
                 ResultSet resultSet = preparedStatement.executeQuery();
 
-                int result = -1;
+                long result = -1;
 
                 while(resultSet.first()) {
-                    result = resultSet.getInt("SUM(country.Population)");
+                    result = resultSet.getLong("SUM(country.Population)");
                 }
                 return result;
             }
